New Feature
Added an argument to the command that allows you to, basically, specify
the lowest faction rank added to the region. For example, if you type
officer, it adds only officers and leader to the region. if you type
member, it adds members and up. If you type all, it allows recruits.